requirements.txt
absl-py==0.13.0
adal==1.2.7
aenum==3.1.0
aiohttp==3.7.4.post0
appdirs==1.4.4
applicationinsights==0.11.10
argcomplete==1.12.3
argon2-cffi==20.1.0
astroid==2.5.6
async-generator==1.10
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==21.2.0
autograd==1.3
azure-cli-core==2.24.2
azure-cli-telemetry==1.0.6
azure-common==1.1.27
azure-core==1.14.0
azure-mgmt-core==1.2.2
azure-nspkg==3.0.2
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
azure-storage-nspkg==3.1.0
backcall==0.2.0
backports.lzma==0.0.14
bcrypt==3.2.0
bidict==0.21.2
bleach==3.3.0
bokeh==2.3.3
boto==2.49.0
boto3==1.17.87
botocore==1.20.87
cachetools==4.2.2
carla==0.9.9
certifi==2021.5.30
cffi==1.14.6
cfgv==3.3.0
chardet==4.0.0
click==8.0.1
cloudpickle==1.6.0
colorama==0.4.4
configparser==5.0.2
control==0.9.0
coverage==5.5
crcmod==1.7
cryptography==3.4.7
cycler==0.10.0
cysignals==1.10.3
Cython==0.29.23
dask==2021.7.2
datadog==0.41.0
decorator==4.4.2
defusedxml==0.7.1
dictdiffer==0.8.1
dictor==0.1.7
dill==0.3.4
distlib==0.3.2
distributed==2021.7.2
docker-pycreds==0.4.0
elasticsearch==7.13.1
entrypoints==0.3
ephem==4.0.0.2
fastcluster==1.2.3
filelock==3.0.12
filterpy==1.4.5
flake8==3.9.2
Flask==2.0.1
Flask-Cors==3.0.10
Flask-SocketIO==5.1.0
flatbuffers==2.0
fsspec==2021.7.0
ft4222==1.2.2
future==0.18.2
GeoAlchemy2==0.8.5
git-pylint-commit-hook==2.5.1
gitdb==4.0.7
GitPython==3.1.18
google-auth==1.35.0
google-auth-oauthlib==0.4.6
greenlet==1.1.0
grpcio==1.40.0
gunicorn==20.1.0
h5py==3.2.1
HeapDict==1.0.1
hexdump==3.3
humanfriendly==9.1
hypothesis==6.13.12
identify==2.2.8
idna==2.10
imageio==2.9.0
imgaug==0.4.0
inputs==0.5
ipykernel==5.5.5
ipython==7.24.1
ipython-genutils==0.2.0
ipywidgets==7.6.3
isodate==0.6.0
isort==5.8.0
itsdangerous==2.0.1
jedi==0.18.0
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
json-logging-py==0.2
json-rpc==1.13.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
Keras-Applications==1.0.8
kiwisolver==1.3.1
knack==0.8.2
lazy-object-proxy==1.6.0
libusb1==1.9.2
locket==0.2.1
lru-dict==1.1.7
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.2
matplotlib-inline==0.1.2
mccabe==0.6.1
mistune==0.8.4
mmcv==1.3.9
-e git+https://github.com/open-mmlab/mmdetection.git@475c6bec197e6b495d48146636301590165d8f66#egg=mmdet
mock==4.0.3
motmetrics==1.2.0
mpld3==0.5.2
mpmath==1.2.1
msal==1.12.0
msgpack==1.0.2
msgpack-python==0.5.6
msrest==0.6.21
msrestazure==0.6.4
multidict==5.1.0
mypy==0.812
mypy-extensions==0.4.3
nbclient==0.5.3
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5.1
nodeenv==1.6.0
nose==1.3.7
notebook==6.4.0
numpy==1.21.0
oauthlib==3.1.1
onnx==1.9.0
onnxruntime==1.8.0
onnxruntime-gpu==1.8.1
opencv-python==4.5.2.52
osmium==3.1.3
packaging==20.9
pandas==1.2.4
pandocfilters==1.4.3
parameterized==0.8.1
paramiko==2.7.2
parso==0.8.2
partd==1.2.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.2.0
pipenv==2020.8.13
pkginfo==1.7.0
portalocker==1.7.1
pprofile==2.0.5
pre-commit==2.13.0
prometheus-client==0.11.0
promise==2.3
prompt-toolkit==3.0.18
protobuf==3.17.2
psutil==5.8.0
ptyprocess==0.7.0
py-trees==0.8.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycapnp==1.1.0
pycocotools==2.0.2
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.10.1
pycurl==7.43.0.6
pyDeprecate==0.3.1
pydot==1.4.2
pyflakes==2.3.1
pygame==2.0.0.dev8
Pygments==2.9.0
PyJWT==2.1.0
pylint==2.8.3
pymongo==3.11.4
PyMySQL==0.10.1
PyNaCl==1.4.0
pynmea2==1.18.0
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyprof2calltree==1.4.5
pyproj==3.1.0
pyrsistent==0.17.3
pyserial==3.5
pytest-benchmark==3.4.1
python-dateutil==2.8.1
python-engineio==4.2.0
python-logstash==0.4.6
python-socketio==5.3.0
pytorch-lightning==0.8.5
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.1.0
-e git+https://github.com/mageofboy/qdtrack.git@3458dffffb9a075852bf9d480f88a8e6f70ba9bf#egg=qdtrack
qtconsole==5.1.0
QtPy==1.9.0
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
reverse-geocoder==1.5.1
rsa==4.7.2
s2sphere==0.2.5
s3transfer==0.4.2
scikit-image==0.18.1
scikit-learn==0.24.2
scipy==1.6.3
SCons==4.1.0.post1
seaborn==0.11.1
Send2Trash==1.5.0
sentry-sdk==1.3.0
setproctitle==1.2.2
Shapely==1.7.1
shortuuid==1.0.1
simplejson==3.17.2
six==1.16.0
smbus2==0.4.1
smmap==4.0.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.17
subprocess32==3.5.4
sympy==1.8
tabulate==0.8.9
tblib==1.7.0
tenacity==7.0.0
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
terminado==0.10.0
terminaltables==3.1.0
testpath==0.5.0
threadpoolctl==2.2.0
tifffile==2021.4.8
timezonefinder==5.2.0
toml==0.10.2
toolz==0.11.1
torch==1.9.0+cu111
torchaudio==0.9.0
torchmetrics==0.5.1
torchvision==0.10.0+cu111
tornado==6.1
tqdm==4.61.0
traitlets==5.0.5
typed-ast==1.4.3
typing-extensions==3.10.0.0
urllib3==1.26.5
utm==0.7.0
virtualenv==20.4.7
virtualenv-clone==0.5.4
wandb==0.12.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.0.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xmltodict==0.12.0
yarl==1.6.3
zict==2.0.0
zmq==0.0.0
